Adding in VhdFile support, part of fixes #6 (#17)
This pull request includes significant updates to the Pulumi Hyper-V
provider, adding new functionality and improving existing features. The
most important changes include the addition of a new `VhdFile` resource,
enhancements to the `Machine` resource, and updates to the provider's
schema and documentation.

### New Resource Addition:
* Added new `VhdFile` resource to manage VHD (Virtual Hard Disk) files
for Hyper-V virtual machines. This includes properties for `path`,
`diskType`, `sizeBytes`, and lifecycle methods such as `Create`, `Read`,
`Update`, and `Delete`.
[[1]](diffhunk://#diff-900b143e9b41f285edd03b0f19fa2d9a3de8ffd7757b2676d62e34ce3d9e6a5cR156-R245)
[[2]](diffhunk://#diff-418e6e7d93aeacebe7c22d5fb674b48bfa15c6a514eb3ae4744b86d447595d8eR23)
[[3]](diffhunk://#diff-418e6e7d93aeacebe7c22d5fb674b48bfa15c6a514eb3ae4744b86d447595d8eR95-R99)
[[4]](diffhunk://#diff-69bf3006e5956d30bcffd2eb3e0a0f15fe9666b440e63b53db2f1f7c455304dcL1-L29)

### Enhancements to Machine Resource:
* Introduced `triggers` property to the `Machine` resource, enabling
resource replacement when specified values change. This ensures that
changes in the `triggers` array result in the resource being recreated
rather than updated in place.
[[1]](diffhunk://#diff-900b143e9b41f285edd03b0f19fa2d9a3de8ffd7757b2676d62e34ce3d9e6a5cL72-R72)
[[2]](diffhunk://#diff-bc668ae361f8ae8f2241c0c0c0c7c7edfc0e06786100606ed903e61a5c1bb5c1R46-R71)
* Updated the `Create` method to set Secure Boot to false for Generation
1 VMs, as they do not support Secure Boot.

### Documentation Updates:
* Corrected the documentation link for the Pulumi Hyper-V Provider.

These changes enhance the functionality and usability of the Pulumi
Hyper-V provider, making it more robust and feature-rich for managing
Hyper-V resources.