chore(testing): re-enabling esc via .ci-mgmt.yaml (#36)
This pull request includes changes to the CI/CD configuration and
workflow files to re-enable and update the test jobs for Windows. The
most important changes are summarized below:

Re-enabling and updating test jobs:

*
[`.github/workflows/run-acceptance-tests-windows.yml`](diffhunk://#diff-f095ef392962db34a7bf6e6e53683851edd60474ccc72732562ff0ce6dc043d3L267-R392):
Re-enabled the `test` job and updated its configuration to run on
`windows-latest` with a matrix of languages including `nodejs`,
`python`, `dotnet`, `go`, and `java`. This includes steps for setting up
the environment, installing dependencies, and running tests.
[[1]](diffhunk://#diff-f095ef392962db34a7bf6e6e53683851edd60474ccc72732562ff0ce6dc043d3L267-R392)
[[2]](diffhunk://#diff-f095ef392962db34a7bf6e6e53683851edd60474ccc72732562ff0ce6dc043d3L409-R409)

Configuration updates:

*
[`.ci-mgmt.yaml`](diffhunk://#diff-b14f05d67d5776f50ea7fd6642fb7f6f5d765e1acf0b9827928b6081d201b1baR1):
Added a configuration to disable AWS in the CI management settings.