chore: Adding experimental note (#41)
This pull request includes a small but important change to the
`README.md` file. The change adds a warning message indicating that the
project is experimental and currently non-functional between commits,
with no support available.

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R9-R10):
Added a warning message to indicate that the project is experimental,
non-functional between commits, and unsupported.